### 2024-09-23 #7267 @danielskatz
👋 @HPicatto - thanks for your submission.

Before we proceed, there are a few items that need updating:

1. For the affiliations, we don't really need addresses, just institutions and countries.
2. Your paper is too long. JOSS requests papers that are roughly 250-1000 words, and yours is over twice as long. See the [example paper](https://joss.readthedocs.io/en/latest/example_paper.html). Perhaps there is material you can remove and either move to the README or elsewhere in the repo, or in the documentation? (For example, the implementation challenges section might fit a traditional paper about the software, but it doesn't really fit a JOSS paper. And the platform comparison part might be moved to the repo.) Feel free to use the command `@editorialbot check repository` to run some checks, one of which calculates the word count of the paper. editorialbot commands need to be the first entry in a new comment.

Once these changes are made, please ping me and we can get the review started.
