### 2024-09-24 9d80fbb
Information measurement (#93)
* explore information of landing clustering, entropy

* explore (draft) gpu based processing and ML

---------

Co-authored-by: Hernan Picatto <hernan.picatto@ascii.ac.at>
