+++ src/pipelines/ascii_library/ascii_library/orchestration/pipes/instance_config.py
@@ -1,9 +1,11 @@
 import json
 import re
-from typing import Any, Dict, Optional, Sequence
+from typing import Any, Dict, List, Optional, Sequence
 
 from ascii_library.orchestration.resources import emr_constants
-from boto3 import client
+from mypy_boto3_emr import EMRClient
+from mypy_boto3_emr.type_defs import ListSupportedInstanceTypesOutputTypeDef
+from mypy_boto3_pricing import PricingClient
 
 
 class CloudInstanceConfig:
@@ -328,32 +330,32 @@ class CloudInstanceConfig:
 
     def get_available_instances(
         self,
-        emr_client: client,
+        emr_client: EMRClient,
         releaseLabel: Optional[str] = emr_constants.releaseLabel,
         **kwargs,
-    ):
+    ) -> List[Any]:
+        if releaseLabel is None:
+            raise ValueError("releaseLabel cannot be None")
         result = []
-        response: Dict[str, Any] = {}
+        response: Optional[ListSupportedInstanceTypesOutputTypeDef] = None
         while True:
-            if response.get("Marker") is None:
+            if response is None or response.get("Marker") is None:
                 response = emr_client.list_supported_instance_types(
                     ReleaseLabel=releaseLabel
                 )
             else:
+                marker = response["Marker"]
                 response = emr_client.list_supported_instance_types(
-                    ReleaseLabel=releaseLabel, Marker=response.get("Marker")
+                    ReleaseLabel=releaseLabel, Marker=marker
                 )
-            result.extend(
-                self.add_prefix_suffix_to_instances(
-                    response.get("SupportedInstanceTypes")
-                )
-            )
+            supported_instance_types = response.get("SupportedInstanceTypes", [])
+            result.extend(self.add_prefix_suffix_to_instances(supported_instance_types))
             if response.get("Marker") is None:
                 break
         return result
 
     def get_instance_w_price(
-        self, price_client: client, instances: Sequence[Dict], n: int = 15
+        self, price_client: PricingClient, instances: Sequence[Dict], n: int = 15
     ):
         for instance in instances:
             response = price_client.get_products(
@@ -389,7 +391,9 @@ class CloudInstanceConfig:
             instanceTypeConfigs.append(temp)
         return instanceTypeConfigs
 
-    def get_fleet_programatically(self, emrClient: client, priceClient: client):
+    def get_fleet_programatically(
+        self, emrClient: EMRClient, priceClient: PricingClient
+    ):
         filter_values = [
             self.filters.get("VCPU"),
             self.filters.get(
