+++ src/pipelines/ascii_library/ascii_library/orchestration/pipes/emr.py
@@ -19,7 +19,6 @@ from ascii_library.orchestration.resources.emr_constants import pipeline_bucket
 from ascii_library.orchestration.resources.utils import (
     get_dagster_deployment_environment,
 )
-from boto3 import client
 from dagster import get_dagster_logger
 from dagster._annotations import experimental
 from dagster._core.definitions.resource_annotation import ResourceParam
@@ -32,6 +31,10 @@ from dagster._core.pipes.client import (
 )
 from dagster._core.pipes.utils import open_pipes_session
 from dagster_pipes import PipesExtras
+from mypy_boto3_emr import EMRClient
+from mypy_boto3_emr.type_defs import StepConfigTypeDef
+from mypy_boto3_pricing import PricingClient
+from mypy_boto3_s3 import S3Client
 
 
 @experimental
@@ -48,9 +51,9 @@ class _PipesEmrClient(_PipesBaseCloudClient):
 
     def __init__(
         self,
-        emr_client: client,
-        s3_client: client,
-        price_client: client,
+        emr_client: EMRClient,
+        s3_client: S3Client,
+        price_client: PricingClient,
         bucket: str,
         context_injector: Optional[PipesContextInjector] = None,
         message_reader: Optional[PipesMessageReader] = None,
@@ -143,7 +146,7 @@ class _PipesEmrClient(_PipesBaseCloudClient):
         bucket: str,
         s3_path: str,
         emr_job_config: Dict[str, Any],
-        step_config: Dict[str, Any],
+        step_config: StepConfigTypeDef,
         libraries_to_build_and_upload: Optional[List[str]] = None,
         libraries: Optional[List[LibraryConfig]] = None,
         extras: Optional[PipesExtras] = None,
@@ -171,12 +174,72 @@ class _PipesEmrClient(_PipesBaseCloudClient):
             extras["step_config"] = step_config
         return extras, emr_job_config
 
+    def adjust_emr_job_config(
+        self,
+        emr_job_config: dict,
+        fleet_config: Optional[CloudInstanceConfig],
+    ) -> dict:
+        if (
+            emr_job_config["Instances"].get("InstanceGroups") is None
+            and emr_job_config["Instances"].get("InstanceFleets") is None
+        ):
+            if fleet_config is not None:
+                emr_job_config["Instances"]["InstanceFleets"] = (
+                    fleet_config.get_fleet_programatically(
+                        emrClient=self._emr_client, priceClient=self._price_client
+                    )
+                )
+                emr_job_config["ManagedScalingPolicy"]["ComputeLimits"][
+                    "UnitType"
+                ] = "InstanceFleetUnits"
+            else:
+                raise ValueError(
+                    "No instance groups or fleets defined, and fleet_config is None."
+                )
+        return emr_job_config
+
+    def submit_emr_job(
+        self,
+        emr_job_config: dict,
+        step_config: StepConfigTypeDef,
+        bucket: str,
+        extras: PipesExtras,
+    ) -> str:
+        emr_job_config = self.modify_env_var(
+            cluster_config=emr_job_config, key="bucket", value=bucket
+        )
+        emr_job_config = self.modify_env_var(
+            cluster_config=emr_job_config, key="key", value=self._key_prefix
+        )
+
+        job_flow = self._emr_client.run_job_flow(**emr_job_config)
+        get_dagster_logger().debug(f"EMR configuration: {job_flow}")
+        self._emr_client.add_tags(
+            ResourceId=job_flow["JobFlowId"],
+            Tags=[
+                {"Key": "jobId", "Value": job_flow["JobFlowId"]},
+                {"Key": "executionMode", "Value": extras["execution_mode"]},
[…]
