### 2024-09-27 #7267 @HPicatto
Hi @danielskatz,
We have finished resizing the content in an internal branch, and we moved two sections into .md files. Could you guide us on how we should reference these sections within the main Markdown file so that they are properly linked in the paper? Should we use the final GitHub URL, or is there a specific way that JOSS handles appendices or external references?
Thank you for your help!
