### 2024-09-27 #7267 @danielskatz
JOSS does not consider these appendices, but they are external references, so you can just use URLs to them, saying that they are in the GitHub repo
