### 2024-09-30 70a7852
migrate to pixi (#115)
* migrate to pixi
* upgrade dependencies
* change all lining and formatting from mypy black to ruff and pyright
