### 2024-10-04 0df187a
Sync README.md and LICENSE.txt from research-space to main branch of ascii-hydra
