### 2024-10-04 #7267 @danielskatz
@HPicatto - All the commands I'm now going to run to check the length of the paper, check references, and regenerate the paper are all commands you can run too.  Note that editorialbot commands need to be the first entry in a new comment.
