### 2024-10-04 #7267 @danielskatz
@HPicatto - One issue is that the paper still appears to be twice a long as JOSS recommends. As I skim it, the Implementation Challenges section seems to not really fit JOSS's model, though if this was a paper in another venue about the process of developing the software, it would make sense. That's the only section I would suggest removing, and without it, I would be ok going ahead with the review even if the paper was still a bit long.

Another minor issue is that we don't need your mailing addresses in your affiliations - just institution and country is enough, along with unit (e.g. department, division, etc.) if you want.

Thanks for the progress!
