### 2024-10-09 a213597
Paper feedback 2.0 (#160)
* reupload changes

* move sections into appendix

* check how it compiles

* fix path apendix

* fix file name

* update links on appendix 2

* remove address and province

* check with braces

---------

Co-authored-by: Hernan <hernan.picatto@ascii.ac.at>
