### 2024-10-11 #7267 @HPicatto
Hi @danielskatz new version is here, now the extension is 1300 words, do you think it's acceptable? it's not easy to keep reducing the length
