+++ src/pipelines/ascii_library/ascii_library/orchestration/pipes/spark_pipes_factory.py
@@ -5,14 +5,20 @@ from typing import Any, List, Mapping, Optional, Sequence, Union
 from dagster import (
     AssetExecutionContext,
     AssetsDefinition,
+    Config,
     MaterializeResult,
     PartitionsDefinition,
     PipesSubprocessClient,
     asset,
 )
 from databricks.sdk.service import jobs
+from pydantic import Field
 
-from ascii_library.orchestration.pipes import LibraryConfig, LibraryKind
+from ascii_library.orchestration.pipes import (
+    LibraryConfig,
+    LibraryKind,
+    get_engine_by_value,
+)
 from ascii_library.orchestration.pipes.databricks import PipesDatabricksEnhancedClient
 from ascii_library.orchestration.pipes.emr import PipesEmrEnhancedClient
 from ascii_library.orchestration.pipes.instance_config import CloudInstanceConfig
@@ -38,7 +44,8 @@ def get_libs_dict(
                 )
                 libs_dict.append({lib.kind.value: {"package": package_str}})
             elif lib.kind == LibraryKind.Wheel:
-                libs_dict.append({lib.kind.value: lib.name_id})
+                lib_name = f"s3://{pipeline_bucket}/{lib.name_id}"
+                libs_dict.append({lib.kind.value: lib_name})
 
     return libs_dict
 
@@ -63,6 +70,28 @@ def generate_uploaded_script_paths(
     return new_path
 
 
+def get_engine_from_config(config, spark_pipes_client):
+    if "config" in config.keys():
+        if (
+            "override_default_engine" in config["config"].keys()
+            and config["config"]["override_default_engine"]
+        ):
+            override_engine = get_engine_by_value(
+                config["config"]["override_default_engine"]
+            )
+            return spark_pipes_client.get_spark_pipes_client(
+                override_engine
+            ), override_engine
+        else:
+            return spark_pipes_client.get_spark_pipes_client(
+                spark_pipes_client.engine
+            ), spark_pipes_client.engine
+    else:
+        return spark_pipes_client.get_spark_pipes_client(
+            spark_pipes_client.engine
+        ), spark_pipes_client.engine
+
+
 def spark_pipes_asset_factory(  # noqa: C901
     name: str,
     key_prefix: Sequence[str],
@@ -79,7 +108,6 @@ def spark_pipes_asset_factory(  # noqa: C901
     emr_additional_libraries: Optional[List[LibraryConfig]] = None,
     dbr_additional_libraries: Optional[List[LibraryConfig]] = None,
     emr_job_config: Optional[dict] = None,
-    override_default_engine: Optional[Engine] = None,
     fleet_filters: Optional[CloudInstanceConfig] = None,
     # TODO: in the future support IO manager perhaps for python reading directly from S3 io_manager_key:Optional[str]: None
     # but maybe we intend to keep the offline sync process
@@ -91,16 +119,8 @@ def spark_pipes_asset_factory(  # noqa: C901
     Automatically configure the right pipes clients.
     In the case of Databricks: Ensure dependencies and scripts are present and ready to be executed - automatically build the dependent libraries and upload these to DBFS.
     """
-    client: Union[
-        PipesSubprocessClient,
-        PipesDatabricksEnhancedClient,
-        PipesEmrEnhancedClient,
-    ] = spark_pipes_client.get_spark_pipes_client(override_default_engine)
-    if override_default_engine is not None:
-        engine_to_use = override_default_engine
-    else:
-        engine_to_use = spark_pipes_client.engine
 
+    engine_to_use = spark_pipes_client.engine
[…]
