### 2024-10-21 #7267 @danielskatz
@HPicatto - sorry I missed your note above.  Yes, I think this length is fine.  Can you also add the missing DOI that editorialbot found in your .bib file, then please recheck the references and regenerate the pdf?  I will mark this as unpaused, but will need to wait until I can find an editor with capacity to take this on.
