+++ papers/multi_platform_orchestration/paper.bib
@@ -85,7 +85,9 @@ series = {SIGMOD '22}
   number={3},
   pages={2011--2041},
   year={2020},
-  publisher={Springer}
+  publisher={Springer},
+  url = {https://doi.org/10.1007/s11192-020-03726-9},
+  doi = {10.1007/s11192-020-03726-9},
 }
 @article{kumar,
   title={Vendor Lock-In Situation and Threats in Cloud Computing},
