### 2024-11-07 #7267 @rico-berner
> hi @d-eggert @rico-berner 👋 would you be interested in and available to review this JOSS submission? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: [joss.readthedocs.io/en/latest/review_criteria.html](https://joss.readthedocs.io/en/latest/review_criteria.html)
>
> If not, could you recommend any potential reviewers? I was hoping to have your insights because of your past authorship of related JOSS publications.

I am sorry, as well, I haven't any capacity to review the paper.
