### 2024-11-08 8d1af0f
add models directory (#195)
* add models directory

---------

Co-authored-by: Georg Heiler <georg.heiler@ascii.ac.at>
