+++ src/pipelines/ascii_library_demo/pipeline_example/pixi.toml
@@ -26,15 +26,15 @@ pyspark = "~=3.5.1"
 openjdk = "~=17.0.11"
 fsspec = ">=2024.6.1"
 s3fs = ">=2024.6.1"
-dagster = "~=1.8.10"
-dagster-webserver = "~=1.8.10"
-dagster-pipes = "~=1.8.10"
+dagster = ">=1.9.0,<1.10"
+dagster-webserver = ">=1.9.0,<1.10"
+dagster-pipes = ">=1.9.0,<1.10"
 
 [feature.example.pypi-dependencies]
 tenacity = ">=8.5.0,<10"
-dagster-aws = ">=0.24.4,<0.25"
-dagster-pyspark = ">=0.24.4,<0.25"
-dagster-databricks = ">=0.24.4,<0.25"
+dagster-aws = ">=0.25.0,<0.26"
+dagster-pyspark = ">=0.25.0,<0.26"
+dagster-databricks = ">=0.25.0,<0.26"
 
 delta-spark = "~=3.2.0"
 deltalake = ">=0.18.2,<0.19"
