+++ taplo.toml
@@ -1,1 +1,0 @@
-exclude = [".trunk/*", ".pixi/*"]
