### 2024-11-12 e135e8a
fix function (#204)
* fix function

* fmt

* add frozen

---------

Co-authored-by: Georg Heiler <georg.heiler@ascii.ac.at>
