### 2024-12-05 eace3fc
Feat/upgrade dbt duckdb (#218)
* upgrade

* upgrade buenavista

* more updates

* upgrade

---------

Co-authored-by: Georg Heiler <georg.heiler@ascii.ac.at>
