### 2024-12-06 #7267 @HPicatto
Hi @danielskatz do you have any idea on how is the normal process for this kind of papers?
thanks
