### 2024-12-06 df9929f
Change in boto was breaking pipeline (#221)
* add gitignore

* fix missing constant

* clean log with infinite builds

* fmt and lint

* try

* restore old utils

---------

Co-authored-by: Hernan Picatto <hernan.picatto@ascii.ac.at>
Co-authored-by: Georg Heiler <georg.heiler@ascii.ac.at>
