### 2024-12-06 2e8bc58
Feature/config fleet (#223)
* updates node config

* change spot price

* fmt and update

* not fix lint

* fmt

* other file fmt

---------

Co-authored-by: Hernan Picatto <hernan.picatto@ascii.ac.at>
Co-authored-by: Georg Heiler <georg.heiler@ascii.ac.at>
