### 2024-12-07 6c8b7ca
refine bootstrap (#228)
* refine bootstrap

* pin tldxtract

* revert

* use pip3

* revert + pip3

---------

Co-authored-by: Georg Heiler <georg.heiler@ascii.ac.at>
