### 2024-12-08 a8d45c2
fix race conditon on bootstrap (#230)
Co-authored-by: Georg Heiler <georg.heiler@ascii.ac.at>
