### 2024-12-18 ab52a63
issue #240 (#234)
* make it all on demand

* update configs

* implement changes on network

* make part spot

* fix lockfile

* fix formatting

* try ignore

* fix

---------

Co-authored-by: Hernan Picatto <hernan.picatto@ascii.ac.at>
Co-authored-by: Georg Heiler <georg.heiler@ascii.ac.at>
