+++ src/pipelines/ascii_library/ascii_library/orchestration/resources/emr_constants.py
@@ -90,6 +90,15 @@ class VolumeType(Enum):
     magnetic = "standard"
 
 
+class Subnets(Enum):
+    usEast1b = "subnet-02035452bafe54090"  # main
+    usEast1e = "subnet-04be55a5e7990e7fe"
+    usEast1d = "subnet-0deb46a6dcbcb90be"
+    usEast1a = "subnet-01c57c11d6c92a586"
+    usEast1c = "subnet-087f4559eec6caddb"
+    usEast1f = "subnet-01e0cb6f08abb93ee"
+
+
 pipeline_bucket = "ascii-supply-chain-research-pipeline"
 mock = False
 sizeInGB = 300
@@ -98,7 +107,7 @@ volumesPerInstance = 1
 ebsOptimized = True
 percentageOfOnDemandPrice = 70.0
 allocationStrategy = AllocationStrategy.priceCapacity
-timeoutDuration = 60
+timeoutDuration = 10
 weightedCapacity = (
     4  # this should be a number greater than 1 and should match the instance vcore
 )
