### 2024-12-20 #7267 @danielskatz
@picattoh - you can certainly suggest people to @HaoZeke (see https://github.com/openjournals/joss-reviews/issues/7267#issue-2541426496)
