### 2025-01-14 #7267 @danielskatz
> What's the criteria for the reviewer?

* Sufficiently experienced in open source to review the software and the paper against the JOSS review criteria (see https://joss.readthedocs.io/en/latest/review_criteria.html)
* not conflicted with the submission (see https://joss.readthedocs.io/en/latest/reviewer_guidelines.html#joss-conflict-of-interest-policy)
* is generally responsive, doesn't drop activities, can provide constructive feedback
