### 2025-01-18 #7267 @HaoZeke
@danielskatz and @HPicatto many many apologies for the delay. I had unexpected travel delays. I'll get started right away.

@Midnighter thank you for volunteering! I'll add you as a reviewer, and then we can get started.
