### 2025-01-18 #7695 @editorialbot
[REVIEW]: Cost-Effective Big Data Orchestration Using Dagster: A Multi-Platform Approach
**Submitting author:** @HPicatto (<a href="http://orcid.org/0000-0001-6662-464X">Hernan Picatto</a>)
**Repository:** https://github.com/ascii-supply-networks/ascii-hydra
**Branch with paper.md** (empty if default branch): main
**Version:** v1.0.0
