### 2025-02-04 ascii-supply-networks/ascii-hydra#8 @abhishektiwari
[JOSS] ASCII hydra - Documentation Improvements
1.  Explanation link in `README.md` is broken https://georgheiler.com/2024/06/21/cost-efficient-alternative-to-databricks-lock-in/.
2. Expand **preprequisites** for `pyspark` local installation. Currently no other documentation is provided so I am not sure how authors of `ascii-hydra` expect a user who may not be familiar the details of to figure out.
3. Is Dagster a prequisite? If yes please provide guidance.
4. What are OS/hardware considerations of running `ascii-hydra` on a local machine? Does it support Mac (ARM vs. Intel), Windows, Linux?

[For JOSS Review](https://github.com/openjournals/joss-reviews/issues/7695)
