### 2025-02-04 ascii-supply-networks/ascii-hydra#9 @abhishektiwari
[JOSS] ASCII hydra - Provide a Dockerfile/Docker Image
To reduce the overall pain to review and use `ascii-hydra` software which poorly documented I strongly recommend authors to provide a Dockerfile/Docker Image with all prerequisite installed. As reviewer I can not use hours and days try to get this software working on my local environment.

[For JOSS Review](https://github.com/openjournals/joss-reviews/issues/7695)
