### 2025-02-04 ascii-supply-networks/ascii-hydra#10 @abhishektiwari
[JOSS] ASCII hydra - Provide PIP Installation
Provide PIP Installation support for `ascii-hydra` so that a user can do something like following to install everything required to run `ascii-hydra`

```
pip install dagster dagster-webserver pyspark ascii-hydra
```

[For JOSS Review](https://github.com/openjournals/joss-reviews/issues/7695)
