### 2025-02-04 ascii-supply-networks/ascii-hydra#11 @abhishektiwari
[JOSS] ASCII hydra - Add Automated Tests and Github Action to run it
There are some tests under `ascii_library_tests` but not clear what is required to run these tests. Please add Github Action to run these tests.  Integration or e2e tests are missing which is essential for a data pipeline package. Bare minimum, I suggest authors to automate the `ascii_library_demo` as e2e test which can be run on every commit or on-demand using a Github actions. Authors are more than welcome to add any additional manual or automated testing. So far I don't think this software meets the quality bar of JOSS and needs to improve.

[For JOSS Review](https://github.com/openjournals/joss-reviews/issues/7695)
