### 2025-02-04 ascii-supply-networks/ascii-hydra#12 @abhishektiwari
[JOSS] API/Package Documentation
Given `ascii_library` is a Python package, strongly recommend authors to provide package API / functionality documentation which can be generated automatically using something like `pydoc` and can be hosted on `Read the Docs`.

[For JOSS Review](https://github.com/openjournals/joss-reviews/issues/7695)
