### 2025-02-04 #7695 @abhishektiwari
@HPicatto - Added a few tickets to improve the documentation and usability of the ASCII Hydra. Appreciate of some of the issues are fixed before take any further review of this package.

@HPicatto - I see multiple people contributing to project but they are not included as authors? Can you clarify if other contributors are aware of this paper? @HaoZeke worth confirming with EiC guideline here.
