### 2025-02-04 ascii-supply-networks/ascii-hydra#13 @abhishektiwari
[JOSS] ASCII hydra - Add contribution/community guidelines
Add contribution/community guidelines

[For JOSS Review](https://github.com/openjournals/joss-reviews/issues/7695)
