### 2025-02-04 ascii-supply-networks/ascii-hydra#14 @abhishektiwari
[JOSS] ASCII hydra - Add a statement of need on README or Documentation Landing Page
Add a statement of need to README.me or Documentation Landing Page (if there is one)

[For JOSS Review](https://github.com/openjournals/joss-reviews/issues/7695)
