### 2025-02-06 #7695 @HPicatto
This paper is part of an institution, for releasing it we are using a github action using [gitexporter](https://github.com/open-condo-software/gitexporter). It was mentioned in this [issue ](https://github.com/openjournals/joss/issues/1339) of JOSS. The other collaborators are users of this software and are aware of the existance and submission of the paper. They only collaborated with research that is not published/publishable. They also did not contribute in any meaningful way to the software or the paper, the coding was done by Georg and myself.
