### 2025-02-06 ascii-supply-networks/ascii-hydra#9 @HPicatto
We will try to make a docker, but are you getting any particular error? the [readme](https://github.com/ascii-supply-networks/ascii-hydra?tab=readme-ov-file#creation-of-environment) explains that the only thing that you should run is `pixi run start` and it should work in any OS (we have tested it in Windows and regularly use it on Linux and MacOS)
