### 2025-02-06 ascii-supply-networks/ascii-hydra#8 @HPicatto
Hi, I'm going to update the link to Georg link. Thanks for mentioning this! I think that this is a bug on the syncronization between the private repo and this public one. you should NOT install anything by yourself, there it should work out of the box when running pixi start, but in the syncronization we at some point we remove the `pixi.toml`. I will modify the action tomorrow and ping you back for testing
