### 2025-02-06 ascii-supply-networks/ascii-hydra#10 @HPicatto
same than #8 , you should be able to use it running `pixi install` but there is a bug
