### 2025-02-07 ascii-supply-networks/ascii-hydra#10 @geoHeil
You do not even eed to run pixi install - a `pixi run mytask` will bootstrap everything for you smoothly.
