### 2025-02-13 b69156a
Feat/ddb 1.2 (#245)
* prepare update to 1.2.0

* 1.2.0 update

* update files revert to fail more properly

---------

Co-authored-by: Georg Heiler <georg.heiler@ascii.ac.at>
