### 2025-02-14 cfe4e06
deps upgrade (#249)
* deps upgrade

* fix

---------

Co-authored-by: Georg Heiler <georg.heiler@ascii.ac.at>
