+++ src/pipelines/ascii_library/pyproject.toml
@@ -11,7 +11,7 @@ authors = [
   { name = "Hernan Picatto", email = "hernan.picatto@ascii.ac.at" },
 ]
 license = { text = "MIT" } # Or another license identifier
-dependencies = ["dagster>=1.9.0,<1.10", "dagster-pipes>=1.9.0,<1.10", "boto3"]
+dependencies = ["dagster>=1.10.0,<1.11", "dagster-pipes>=1.10.0,<1.11", "boto3"]
 
 [project.optional-dependencies]
 
