### 2025-03-03 a83a8df
update sync and pixi.toml
