+++ src/pipelines/ascii_library_demo/pipeline_example/pixi.toml
@@ -1,5 +1,5 @@
 [dependencies]
-python = "3.11"
+python = "~=3.11"
 pip = ">=24.2"
 
 [environments]
@@ -8,38 +8,40 @@ example = { features = ["example"], solve-group = "example" }
 testing = { features = ["ci", "example"], solve-group = "example" }
 
 [feature.ci.dependencies]
-pytest = ">=8.3.2,<9"
+pytest = ">=8.3.3,<9"
 pytest-mock = ">=3.14.0,<4"
 pytest-cov = ">=5.0.0,<6"
-pyright = "~=1.1.375"
+pyright = "~=1.1.388"
 ruff = ">=0.5.7,<1"
+yamllint = ">=1.35.1,<2"
 taplo = ">=0.9.3,<0.10"
+nodejs = ">=20.8.1,<23"
 
 [feature.ci.pypi-dependencies]
-sqlfluff = "~=3.1.0"
-moto = "~=5.0.12"
+sqlfluff = "~=3.2.0"
+moto = "~=5.0.14"
 
 [feature.example.dependencies]
 grpcio = "*"
 grpcio-tools = "*"
-pyspark = "~=3.5.1"
+pyspark = "~=3.5.3"
 openjdk = "~=17.0.11"
-fsspec = ">=2024.6.1"
-s3fs = ">=2024.6.1"
+fsspec = ">=2024.9.0"
+s3fs = ">=2024.9.0"
 dagster = ">=1.10.0,<1.11"
 dagster-webserver = ">=1.10.0,<1.11"
 dagster-pipes = ">=1.10.0,<1.11"
 
 [feature.example.pypi-dependencies]
-tenacity = ">=8.5.0,<10"
+tenacity = ">=9.0.0,<10"
 dagster-aws = ">=0.26.0,<0.27"
 dagster-pyspark = ">=0.26.0,<0.27"
 dagster-databricks = ">=0.26.0,<0.27"
 
-delta-spark = "~=3.2.0"
-deltalake = ">=0.18.2,<0.19"
-build = "~=1.2.1"
-databricks-sdk = ">=0.8,<0.31"
+delta-spark = "~=3.2.1"
+deltalake = ">=0.22.3,<0.23"
+build = "~=1.2.2"
+databricks-sdk = ">=0.8,<0.33"
 ascii_library = { path = "../../ascii_library", editable = true }
 pipeline_example = { path = "./", editable = true }
 
