### 2025-03-03 cd1eedc
fix running spark
