### 2025-03-04 75e831d
fix steps so it doesn't rely on any foreign data
