### 2025-03-05 a6adcb2
move files from ascii_library_demo into base src and sync from research-space to main branch of ascii-hydra
