### 2025-03-06 f01d105
Feature/llm exploration (#241)
* refactor models

* fmt lint

* work on filter LDA

* make clustering store the model

* fix bug on emr

* working on pyspark

* lint fmt

* logging wandb

* remove udf from tokenization

* fmt and lint, cluster working without udf

* fmt and lint

* results from test

* add evoc

* last exploration

* analysis

* precomit

* wrong lock

* refactor

* llm

* update gode

* also this

* make plot work

* fmt and lint

* last version evoc

* add fastopic

* remove output on notebooks

* make pixi toml versions same as main

* fmt

* update docker

* update pixi

*
[…]
