### 2025-03-06 3a16971
Paper/review (#254)
* update sync and pixi.toml

* remove experimental that was causing error

* fix running spark

* change cp to mv

* fix problem on example

* fmt

* updated pixi

* fix steps so it doesn't rely on any foreign data

* add an script to explain how to use pipeline

* fix url, and add contributing

* finish contribution

* update sync

* update path

* fix path

* create path

* check new path

* disable branches

---------

Co-authored-by: Hernan Picatto <hernan.picatto@ascii.ac.at>
