+++ src/pipelines/ascii_library/ascii_library/orchestration/resources/spark.py
@@ -30,8 +30,8 @@ def dev_spark_config(
         secret = "NO_SECRET_PROVIDED"
     results = {
         "spark_conf": {
-            "spark.master": "local[7]",
-            "spark.driver.memory": "16G",
+            "spark.master": "local[*]",
+            "spark.driver.memory": os.environ.get("SPARK_DRIVER_MEMORY", "16G"),
             "spark.sql.session.timeZone": "UTC",
             "spark.driver.extraJavaOptions": "-Duser.timezone=UTC",
             "spark.sql.adaptive.enabled": "true",
@@ -52,7 +52,9 @@ def dev_spark_config(
             "spark.hadoop.fs.s3a.path.style.access": "true",
             "spark.hadoop.fs.s3a.impl": "org.apache.hadoop.fs.s3a.S3AFileSystem",
             # "spark.hadoop.fs.s3a.endpoint": url,
-            "spark.hadoop.fs.s3a.buffer.dir": "/data/raid5/data/sparktmp",
+            "spark.hadoop.fs.s3a.buffer.dir": os.environ.get(
+                "SPARK_DIR", "/data/raid5/data/sparktmp"
+            ),
             "spark.hadoop.fs.s3a.fast.upload.buffer": "bytebuffer",
             "spark.hadoop.fs.s3a.fast.upload.active.blocks": "4",
             # "spark.databricks.delta.schema.autoMerge.enabled": "true",
@@ -65,7 +67,7 @@ def dev_spark_config(
             "spark.jars.packages": "io.delta:delta-spark_2.12:3.2.0,org.postgresql:postgresql:42.7.4,org.apache.hadoop:hadoop-aws:3.3.4,org.apache.spark:spark-hadoop-cloud_2.12:3.5.0,com.johnsnowlabs.nlp:spark-nlp_2.12:5.5.0",
             "spark.databricks.delta.schema.autoMerge.enabled": "True",
             "spark.databricks.delta.schema.autoMerge.enabledOnWrite": "True",
-            "spark.local.dir": "/data/raid5/data/sparktmp",
+            "spark.local.dir": os.environ.get("SPARK_DIR", "/data/raid5/data/sparktmp"),
         }
     }
     if url:
