+++ src/pipelines/ascii_library/ascii_library/orchestration/pipes/spark_script_abc.py
@@ -36,17 +36,9 @@ class SparkScriptPipes(ABC):
         if execution_mode == ExecutionMode.SmallDevSampleS3:
             return f"s3a://{self.sample_data}/ascii_seeds"
         elif execution_mode == ExecutionMode.SmallDevSampleLocal:
-            return (
-                Path(
-                    "../../../reference-data/pipeline_sample_data/commoncrawl/ascii_seeds"
-                )
-                .resolve()
-                .as_uri()
-            )
+            return Path(self.reference_data_path, "ascii_seeds").resolve().as_uri()
         elif execution_mode == ExecutionMode.Full:
             return f"s3a://{self.bucket_seed_nodes}/ascii_seeds"
-        else:
-            raise ValueError(f"Invalid Execution mode: {execution_mode.value}")
 
     def get_base_path_IO(
         self,
@@ -58,8 +50,6 @@ class SparkScriptPipes(ABC):
             return Path("z_state/ascii_dev_pipeline").resolve().as_uri()
         elif execution_mode == ExecutionMode.Full:
             return f"s3a://{self.bucket_cc_results}"
-        else:
-            raise ValueError(f"Invalid Execution mode: {execution_mode.value}")
 
     def get_base_path_commoncrawl(
         self,
@@ -75,8 +65,6 @@ class SparkScriptPipes(ABC):
             )
         elif execution_mode == ExecutionMode.Full:
             return "s3a://commoncrawl/"
-        else:
-            raise ValueError(f"Invalid Execution mode: {execution_mode.value}")
 
     def __init__(self):
         engine = Engine(os.environ.get("SPARK_PIPES_ENGINE", "pyspark"))
@@ -91,10 +79,22 @@ class SparkScriptPipes(ABC):
             context.log.info(f"Partition key: {partition_key}")
             context.log.info(f"Execution mode: {execution_mode}")
 
-            self.bucket_seed_nodes = "ascii-supply-chain-research-input"
-            self.bucket_cc_results = "ascii-supply-chain-research-results"
-            self.bucket_cc_dev_results = "ascii-supply-chain-research-dev-results"
-            self.sample_data = "ascii-supply-chain-research-sample-data"
+            self.bucket_seed_nodes = os.environ.get(
+                "BUCKET_SEED_NODES", "default-seed-nodes"
+            )
+            self.bucket_cc_results = os.environ.get(
+                "BUCKET_CC_RESULTS", "default-cc-results"
+            )
+            self.bucket_cc_dev_results = os.environ.get(
+                "BUCKET_CC_DEV_RESULTS", "default-cc-dev-results"
+            )
+            self.sample_data = os.environ.get(
+                "SAMPLE_DATA_BUCKET", "default-sample-data"
+            )
+            self.reference_data_path = os.environ.get(
+                "REFERENCE_DATA_PATH",
+                "../../../reference-data/pipeline_sample_data/commoncrawl",
+            )
 
             if engine == Engine.Local:
                 local_spark_config = context.get_extra("local_spark_config")
