### 2025-03-06 537e7b6
move files from ascii_library_demo into base src and sync from research-space to main branch of ascii-hydra
