### 2025-03-22 ascii-supply-networks/ascii-hydra#9 @abhishektiwari
@HPicatto you are making a lot of assumptions about potential users of Hydra and their understanding of how it is suppose to work for them. I tried to use `pixi` as documented on README and unless I use my own initiative to clone the repository before running `pixi run start`, it goes nowhere. Then there are unexpected dependency issues (see below). Please update README.md with detailed instructions.

```
Error:   × Unable to solve pypi dependencies for the example solve group because no compatible
  │ python interpreter can be installed for the current platform
  help: Try converting your [pypi-dependencies] to conda [dependencies]
```
