### 2025-03-22 #7695 @abhishektiwari
@HaoZeke I am currently blocked and cannot progress with review due to installation issues. Details on ticket.

## Functionality
Unable to validate the functionality of software, unless installation issues are resolved.

## Documentation
A few open tickets to fix the documentation including statement of need, installation instructions, functionality documentation, community guidelines.

 ## Software paper
- Broken link in PDF paper to Appendix_1.md and Appendix_2.md. Please use absolute URLs and not relative paths.

@HPicatto - Please ping back when these issues are installation instructions are improved.
