### 2025-04-03 8aceda0
update chat UI version (#258)
* update chat UI version

* update

* add volume mounts as RO for tool

* add

* update to duckdb 1.2.1 for neater UI access

* begin exploring tools

* trying - so far failing to add

* update

* cleanup

* add ddb

* add new

* cleanup

* explore gemma

* add vertex 2.5

* refine models

* fix import

* ignore private imports

* fix lint

* fix tests

* i'm an idiot that didn't run fmt before pushing

* solve pyright stuff

* update

* revert

* test

* go back to assertion error

* ignore typed dict

* k

* fix last test

---------

Co-authored-by: Georg Heiler
[…]
