### 2025-04-04 2c257c2
move files from ascii_library_demo into base src and sync from research-space to main branch of ascii-hydra
