+++ src/pipelines/ascii_library_demo/pipeline_example/pixi.toml
@@ -1,25 +1,26 @@
 [dependencies]
-python = "~=3.11"
+python = ">=3.11, <3.12"
 pip = ">=24.2"
 
 [environments]
 ci = { features = ["ci"], solve-group = "ci" }
-example = { features = ["example"], solve-group = "example" }
-testing = { features = ["ci", "example"], solve-group = "example" }
+example = { features = ["ci", "example"], solve-group = "example" }
 
 [feature.ci.dependencies]
 pytest = ">=8.3.3,<9"
 pytest-mock = ">=3.14.0,<4"
 pytest-cov = ">=5.0.0,<6"
-pyright = "~=1.1.388"
 ruff = ">=0.5.7,<1"
 yamllint = ">=1.35.1,<2"
 taplo = ">=0.9.3,<0.10"
 nodejs = ">=20.8.1,<23"
 
 [feature.ci.pypi-dependencies]
-sqlfluff = "~=3.2.0"
 moto = "~=5.0.14"
+nbqa = "~=1.9.0"
+pandas-stubs = "~=2.2.2.240909"
+pyright = "~=1.1.398"
+mypy-boto3-resourcegroupstaggingapi = "~=1.35.26"
 
 [feature.example.dependencies]
 grpcio = "*"
@@ -34,6 +35,7 @@ dagster-pipes = ">=1.10.0,<1.11"
 
 [feature.example.pypi-dependencies]
 tenacity = ">=9.0.0,<10"
+dagster-shared = ">=0.26.0,<0.27"
 dagster-aws = ">=0.26.0,<0.27"
 dagster-pyspark = ">=0.26.0,<0.27"
 dagster-databricks = ">=0.26.0,<0.27"
@@ -42,8 +44,8 @@ delta-spark = "~=3.2.1"
 deltalake = ">=0.22.3,<0.23"
 build = "~=1.2.2"
 databricks-sdk = ">=0.8,<0.33"
-ascii_library = { path = "../../ascii_library", editable = true }
-pipeline_example = { path = "./", editable = true }
+ascii_library = { path = "src/pipelines/ascii_library", editable = true }
+pipeline_example = { path = "src/pipelines/ascii_library_demo/pipeline_example", editable = true }
 
 
 [project]
@@ -69,7 +71,7 @@ cmd = "pytest --cov --cov-report=term-missing ."
 description = "Unittests"
 
 [feature.example.tasks.start]
-cwd = "./"
+cwd = "src/pipelines/ascii_library_demo/pipeline_example"
 cmd = "dagster dev"
 description = "start dagster webserver UI"
 
