### 2025-04-05 7fddda6
move files from ascii_library_demo into base src and sync from research-space to main branch of ascii-hydra
