+++ taplo.toml
@@ -1,0 +1,1 @@
+exclude = [".trunk/*", ".pixi/*"]
