+++ src/pipelines/ascii_library/ascii_library/orchestration/resources/constants.py
@@ -1,10 +1,0 @@
-
-instance_profile = "INTSANCE_PROFILE"
-s3_log_uri = "S3_LOG_URI"
-master_security_group = "MASTER_SECURITY_GROUP"
-slave_security_group = "SLAVE_SECURITY_GROUO"
-service_role = "SERVICE_ROLE"
-job_role = "JOB_ROLE"
-aws_region = "AWS_REGION"
-rackspace_user = "RACKSPACE_USER"
-
