### 2025-04-05 3d1f5d4
move files from ascii_library_demo into base src and sync from research-space to main branch of ascii-hydra
