### 2025-04-05 #7695 @HPicatto
Hi @HaoZeke, @abhishektiwari I just finished adding modifications. you can check the new readme.
new changes:
- add CONTRIBUTING and LICENSE on main directory
- move pixi toml and lock at main directory
- add a file called docs/use_assets.md that explain how to use the example assets, what are environemtal variables or constants
- we tried to make the README.me on main a little bit more clear

for making it clear:
- once the user runs `pixi run start` a dagster ui will be generated for running assets but it is ONLY locally. in order to use Databricks or EMR the user must provide it's own keys an secrets
- for running the automated test as it says in the README.me you can jus symply run `pixi run -e example test` now there is no need of change folder as it was in the previous version
