### 2025-05-19 #7695 @editorialbot
:wave: @openjournals/csism-eics, this paper is ready to be accepted and published.
