### 2025-05-19 #7695 @danielskatz
👋 @HPicatto - I've checked your paper and archive. I've suggested some changes in https://github.com/ascii-supply-networks/ascii-hydra/pull/18. Please merge this, or  let me know what you disagree with.

Also, please change the metadata in your zenodo archive (which you can do (edit) without making a new deposit) so that the title and authors match those of the paper. You also probably want to change the license in the repo.
