### 2025-05-20 #7695 @danielskatz
But @HPicatto - please do merge my wording/etc changes. We can finish these changes now, and if there are more needed based on the remaining review, do this then.

You also may need to make a new archival deposit if the second review leads to changes in the software.
