### 2025-05-21 #7695 @HPicatto
@danielskatz merged. do we have a second reviewier? or it's @HaoZeke  review that counts as second? I'm not sure if I follow the chain of messages in this issue
