### 2025-05-21 #7695 @danielskatz
> [@danielskatz](https://github.com/danielskatz) merged. do we have a second reviewier?

The second reviewer is @Midnighter. I will let @HaoZeke followup and manage/complete the review process, before I come back for the post-review tasks.
