### 2025-05-26 926d85e
new google and claude models (#282)
Co-authored-by: Georg Heiler <georg.heiler@ascii.ac.at>
