+++ papers/multi_platform_orchestration/paper.bib
@@ -21,7 +21,7 @@ abstract = {The serverless cloud computing execution model presents an opportuni
 }
 @article{Zaharia,
 author = {Zaharia, Matei and Xin, Reynold S. and Wendell, Patrick and Das, Tathagata and Armbrust, Michael and Dave, Ankur and Meng, Xiangrui and Rosen, Josh and Venkataraman, Shivaram and Franklin, Michael J. and Ghodsi, Ali and Gonzalez, Joseph and Shenker, Scott and Stoica, Ion},
-title = {Apache {S}park: a unified engine for big data processing},
+title = {Apache Spark: a unified engine for big data processing},
 year = {2016},
 issue_date = {November 2016},
 publisher = {Association for Computing Machinery},
