+++ papers/multi_platform_orchestration/paper.md
@@ -41,8 +41,8 @@ print-pdf:
 ## Summary
 
 The rapid evolution of big data has amplified the need for robust and efficient data processing.
-Spark-based Platform-as-a-Service (PaaS) options, like Databricks and Amazon EMR, offer strong analytics,
-but at the cost of high operational expenses and vendor lock-in [@kumar].
+Spark-based Platform-as-a-Service (PaaS) options, like Databricks and Amazon EMR, offer strong analytics.
+But at the cost of high operational expenses and vendor lock-in [@kumar].
 Despite being user-friendly, their cost structures and opaque pricing can lead to inefficiencies.
 
 This paper introduces a cost-effective, flexible orchestration framework leveraging Dagster [@dagster].
@@ -52,9 +52,9 @@ We showcase Dagster's power to boost efficiency.
 It enforces coding best practices and reduce costs.
 Our implementation showed a 12% speedup over EMR.
 It cut costs by 40% compared to DBR, saving over 300 euros per pipeline run.
-Our framework supports rapid prototyping and testing,
-which is key for continuous development and efficiency.
-It also promotes a sustainable model for large-scale data processing.
+Our framework supports rapid prototyping and testing.
+This is key for continuous development and efficiency.
+It promotes a sustainable model for large-scale data processing.
 
 ## Statement of Need and Relevance
 
@@ -63,10 +63,10 @@ But, they have vendor lock-in and unpredictable costs [@Zaharia].
 This convenience can lead to inefficient resource use, impacting productivity and increasing expenses.
 
 Our solution uses Dagster's orchestration to integrate diverse Spark environments.
-This reduces reliance on a single provider,
-which mitigates lock-in risks, cuts costs, and promotes best coding practices.
-This in turn boosts productivity by supporting rapidly prototyping on smaller datasets.
-It also cuts costs by optimizing resource use, without sacrificing performance.
+This reduces reliance on a single provider.
+This mitigates lock-in risks, cuts costs, and promotes best coding practices.
+This boosts productivity by rapidly prototyping on smaller datasets.
+It cuts costs by optimizing resource use, without sacrificing performance.
 This approach is vital for organizations seeking agile, scalable, and cost-effective data operations.
 
 Also, this approach ensures consistency across development stages.
@@ -76,19 +76,19 @@ It will foster a collaborative scientific environment.
 Their methods will be as open as their findings.
 
 While data pipeline research is growing, existing works focus on different aspects.
-@Anil emphasize optimizing big data processing:
-using energy-efficient scheduling to reduce consumption and latency in data centers.
-@Daw explore using predictive analytics to automate resource scaling in cloud environments,
-which aims to optimize cost and performance.
+Anil et al. [@Anil] emphasize optimizing big data processing.
+Use energy-efficient scheduling to reduce consumption and latency in data centers.
+Daw et al. [@Daw] explore using predictive analytics to automate resource scaling in cloud environments.
+This aims to optimize cost and performance.
 Our multi-cloud strategy leverages open orchestration tools like Dagster.
 This approach bridges existing gaps, deftly managing data tasks across diverse PaaS.
 
 ## Relevance
 
-Our framework improves reproducibility by centralizing metadata management and standardizing orchestration across diverse environments.
+The proposed framework improves reproducibility by centralizing metadata management and standardizing orchestration across diverse environments.
 This in turn reduces infrastructure complexity and aids in consistently replicating experiments, supporting
 reliable research.
-Notwithstanding the growing interest in data pipelines, authors such as @Anil concentrate on the optimisation of big data processing through sophisticated scheduling techniques that minimise energy consumption and latency.
[…]
