+++ pyproject.toml
@@ -1,47 +1,0 @@
-[tool.dagster]
-module_name = "pipeline_example"
-
-[tool.ruff]
-exclude = [
-  ".git",
-  "__pycache__",
-  "docs/source/conf.py",
-  "old",
-  "build",
-  "dist",
-  ".pixi",
-]
-
-line-length = 88
-
-[tool.ruff.lint]
-ignore = ["E501"]
-select = ["F", "E", "W", "C", "B", "I"]
-
-[tool.ruff.lint.mccabe]
-max-complexity = 5
-
-[tool.pyright]
-include = ["."]
-exclude = ["../ascii_library/build"]
-pythonVersion = "3.11"
-venvPath = ".pixi/envs"
-venv = "example"
-extraPaths = ["../ascii_library"]
-
-reportMissingTypeStubs = false
-reportImportCycles = "error"
-useLibraryCodeForTypes = true
-typeCheckingMode = "basic"
-
-[project]
-name = "pipeline_example"
-version = "0.0.1"
-
-[build-system]
-requires = ["setuptools", "wheel", "setuptools-scm"]
-build-backend = "setuptools.build_meta"
-
-[tool.setuptools.packages.find]
-where = ["."]
-exclude = ["tmp", "pipeline_example_tests"]

[…]
