### 2025-05-31 #7695 @HPicatto
Hi @Midnighter I was wondering if you had any feedback or comment on the paper so that we can move on
