### 2025-06-01 #7695 @Midnighter
Hi all,

Painfully aware and extremely sorry that I've been preventing you from progressing. I have blocked out Thursday (5th) to complete my review.
