### 2025-06-09 #7695 @HPicatto
Hi @Midnighter I was wondering if you could do any progress last Thursday or is any way we can help you if anything is blocking you on the reproduction
