### 2025-06-12 #7695 @Midnighter
Unfortunately, Tuesday was completely ruined by urgent work. I will see what I can do this weekend.
