### 2025-07-21 #7695 @danielskatz
@HaoZeke - Please see if you can help this review complete.
