### 2025-07-22 #7695 @Midnighter
I have concluded my review for now and opened issues on the repository with my comments.

My apologies yet again, but as I'm currently on vacation, I don't have access to my company's Databricks or AWS environments. I was therefore not able to fully verify the functionality advertised. I can propose two approaches to remedy this:

1. Many of my comments relate to missing documentation. @HPicatto you could work on those and after I return (August 4), I can verify the Databricks and AWS functionality.
2. I have read the code, run the example until the output step that wants to store data on S3, and @abhishektiwari has run the examples. I would be okay with putting a check mark on functionality and moving on.
