### 2025-08-06 #7695 @HPicatto
Hi @danielskatz sorry I just saw this message. I'm trying to implement the suggestions. Given that they are all question of style and neither bug nor errors on the paper/documentation was reported it's in a relatively low priority. I've some stuff to deliver this week. @Midnighter do you consider that your comments are imperatives for the publication? if so I will make them a priority next week
