### 2025-08-06 #7695 @Midnighter
Ah yes, that was not clearly categorized.

I see the main achievements of your work as follows:

1. You have established how to execute the same task either on local pyspark, AWS EMR, or databricks.
2. Through trial and error, you have established good configurations for those tasks of your workload.
3. From experience, you orchestrate different tasks on one of the three execution platforms.

Now you aim to publish this work such that others may benefit from it, too. Therefore, I see my comments in the following categories:

### Minimum Requirements (please step in @HaoZeke and @danielskatz if I'm contradicting JOSS policy here)

* Address my comments on the paper https://github.com/ascii-supply-networks/ascii-hydra/issues/24
* Publish the library as a package on PyPI, so that others can use your work for their own tasks https://github.com/ascii-supply-networks/ascii-hydra/issues/10
* Document how others can use your library https://github.com/ascii-supply-networks/ascii-hydra/issues/25 and https://github.com/ascii-supply-networks/ascii-hydra/issues/22

### Suggestions and Improvements

* https://github.com/ascii-supply-networks/ascii-hydra/issues/27
* https://github.com/ascii-supply-networks/ascii-hydra/issues/26
* https://github.com/ascii-supply-networks/ascii-hydra/issues/23
* https://github.com/ascii-supply-networks/ascii-hydra/issues/21
* https://github.com/ascii-supply-networks/ascii-hydra/issues/20
* https://github.com/ascii-supply-networks/ascii-hydra/issues/19
