### 2025-08-11 #7695 @HPicatto
totally reasonable and clear now. I will be working on these 3 high priority issue and ping @Midnighter back
