### 2025-09-06 #7695 @HaoZeke
I think the minimum requirements look good. @HPicatto could you give us an update on this?
