### 2025-09-22 #7695 @HPicatto
Hi, I worked on updating the 3 minimum requeriments that we agreed on, but due to updates on DBR and EMR it took me longer than expected. Just yesterday we got the pipeline working again I hope to sync ASCII Hydra as soon as possible
