### 2025-10-31 #7695 @HPicatto
Hi @Midnighter @danielskatz  did you have time to check the last release? does it satisfy the request?
