### 2025-11-01 #7695 @HaoZeke
@HPicatto I see that several of @Midnighter's comments and issues are not closed yet, and @Midnighter has 2 open checklist items still, once those are completed I'm happy to finalize this review. (sorry for the delay here)
