### 2025-11-01 #7695 @HPicatto
@HaoZeke  I mentioned on sept 22 that the three minimum requirements are implemented.
#10, #24, and #25, are all mentioned that were be solved pushes made on august 11 but for closing the issue I'm waiting to see if this satisfies the reviewer that opened it otherwise it would happened like issue #10 that we closed as as solved and @Midnighter reopened it
