### 2025-11-10 #7695 @HPicatto
@Midnighter could you please confirm if we can close these issues (#10, #24, and #25)
