### 2026-01-26 #7695 @danielskatz
@HPicatto - I've suggested some changes in https://github.com/ascii-supply-networks/ascii-hydra/pull/29 - please merge this, or let me know what you disagree with.

In addition, I'm confused by the paragraph above the Relevance section, which seems to strongly overlap the Relevance section. Could you merge the text in these two places?
