### 2026-02-18 #7695 @HPicatto
Hi @danielskatz  i had some personal issues so I this PR was left non review, I have now merged changes that were there and remove the redundancy.
