### 2026-03-12 #7695 @picattoh
Hi @danielskatz the package was listed on [zenodo ](https://zenodo.org/records/15310252) and on [pypi](https://pypi.org/project/ascii-library/) not sure if anything else is neded
