### 2026-03-12 #7695 @HPicatto
the code on both Zenodo and Pypi should be the same as the last version of the review. In fact it was the publication on pypi one of the condition of the reviewers
