### 2026-03-12 #7695 @editorialbot
:wave: @openjournals/csism-eics, this paper is ready to be accepted and published.
