### 2026-03-12 #7695 @danielskatz
👋 @HPicatto - I found a few small things that I suggest changing - see https://github.com/ascii-supply-networks/ascii-hydra/pull/31
