### 2026-03-12 #7695 @danielskatz
Congratulations to @HPicatto (Hernan Picatto) and co-authors on your JOSS publication!!

JOSS completely depends on volunteers - we couldn't do this without you
